Implemented concurrency map cache
The implementation is based on orcaman/concurrent-map but adapted to the storage interface, satisfying the interface method signatures.

Since this is a concurrency-based implementation we could not rely on the normal iteration order for retrieving the oldest entry of the map in case the cache exeeds the max capacity.

For this reason a separate slice has been introduced for storing the map keys. On each insert/remove into/from the map we are also updating the keys slice.